Make llwasm diamond
