+++ spy/llwasm/base.py
@@ -1,116 +1,29 @@
-"""
-A pythonic wrapper around wasmtime.
-
-It is called 'LL' for two reasons:
-
-  - it exposes a low-level view on the code, compared to other wrappers which
-    are more higher level (e.g., the concept of strings doesn't exist, we only
-    have ints, floats and bytes of memory).
-
-  - it's an unused prefix: other prefixes as "Py", "Wasm", "W" etc. would have
-    been very confusing :)
-"""
-
-import sys
-from typing import Any, Optional, Literal
-from typing_extensions import Self
-from asyncio import Future
-import py.path
-import struct
-
-from pyodide.ffi import run_sync
-from pyodide.code import run_js
-loadModule = run_js("""
-    const loadModule = async (f) => {
-        const res = await import(f.replace("/spy", "."));
-        return res.default;
-    };
-    loadModule
-""")
-
 LLWasmType = Literal[None, 'void *', 'int32_t', 'int16_t']
-# ENGINE = wt.Engine()
 
-class LLWasmModule:
+class LLWasmModuleBase:
     f: py.path.local
-    # mod: wt.Module
-
-    def __init__(self, f: py.path.local, *, make_instance=None) -> None:
-        self.f = f
-        if make_instance is None:
-            # JS function that when called makes an instance of the emscripten module
-            self.make_instance = run_sync(loadModule(str(f.new(ext=".mjs"))))
-        else:
-            self.make_instance = make_instance
-
-    @classmethod
-    async def async_new(cls, f: py.path.local):
-        make_instance = await loadModule(str(f.new(ext=".mjs")))
-        return cls(f, make_instance=make_instance)
 
     def __repr__(self) -> str:
         return f'<LLWasmModule {self.f}>'
 
-
-class HostModule:
-    """
-    Base class for host modules.
-
-    Each host module can provide one or more WASM import, used by link().
-    """
-    ll: 'LLWasmInstance' # this attribute is set by LLWasmInstance.__init__
-
-
-class LLWasmInstance:
-
-    def __init__(self, llmod: LLWasmModule,
-                 hostmods: list[HostModule]=[], *, instance=None) -> None:
-        self.llmod = llmod
-
-        if instance is None:
-            self.instance = run_sync(self._make_instance_promise(llmod, hostmods))
-        else:
-            self.instance = instance
-
-        self.mem = LLWasmMemoryPyodide(self.instance.HEAP8)
-        for hostmod in hostmods:
-            hostmod.ll = self
-
+class LLWasmInstanceBase:
+    def get_export(self, name: str) -> Any:
+        raise NotImplementedError
+    
+    def all_exports(self) -> list[Any]:
+        raise NotImplementedError
+    
     @classmethod
     async def async_new(cls, llmod: LLWasmModule, hostmods: list[HostModule]=[]) -> None:
-        return cls(llmod, hostmods, instance=await cls._make_instance_promise(llmod, hostmods))
-
-    @staticmethod
-    def _make_instance_promise(llmod: LLWasmModule, hostmods: list[HostModule]) -> Future[Any]:
-        def adjust_imports(imports):
-            from js import Object
-            env = imports.env
-            for [name, val] in Object.entries(env):
-                if not val.stub:
-                    continue
-                for hostmod in hostmods:
-                    if x := getattr(hostmod, "env_" + name, None):
-                        setattr(env, name, x)
-                        break
-        return llmod.make_instance(adjustWasmImports=adjust_imports)
+        raise NotImplementedError
 
     @classmethod
     def from_file(cls, f: py.path.local,
                   hostmods: list[HostModule]=[]) -> Self:
-        llmod = LLWasmModule(f)
-        return cls(llmod, hostmods)
-
-    def get_export(self, name: str) -> Any:
-        return getattr(self.instance, "_" + name)
-
-    def all_exports(self) -> Any:
         raise NotImplementedError
-        exports = self.instance(self.store)
-        return list(exports._extern_map)
 
     def call(self, name: str, *args: Any) -> Any:
-        func = self.get_export(name)
-        return func(*args)
+        raise NotImplementedError
 
     def read_global(self, name: str, deref: LLWasmType = None) -> Any:
         r"""
@@ -143,8 +56,6 @@ class LLWasmInstance:
             read_global('a', deref='i16') == 0xAAAA
             read_global('b', deref='i16') == 0xBBBB # first item of the array
         """
-        # raise NotImplementedError
-
         addr = self.get_export(name)
         if deref is None:
             return addr
@@ -156,6 +67,13 @@ class LLWasmInstance:
             assert False, f'Unknown type: {deref}'
 
 
+class HostModule:
+    """
+    Base class for host modules.
+
+    Each host module can provide one or more WASM import, used by link().
+    """
+    ll: 'LLWasmInstanceBase' # this attribute is set by LLWasmInstance.__init__
 
 
 class LLWasmMemoryBase:
@@ -214,17 +132,3 @@ class LLWasmMemoryBase:
         """
         self.write_i32(addr, v_addr)
         self.write_i32(addr+4, v_length)
-
-
-class LLWasmMemoryPyodide(LLWasmMemoryBase):
-    def __init__(self, jsmem):
-        self.jsmem = jsmem
-
-    def read(self, addr: int, n: int) -> bytearray:
-        """
-        Read n bytes of memory at the given address.
-        """
-        return self.jsmem.subarray(addr, addr+n).to_py()
-
-    def write(self, addr: int, b: bytes) -> None:
-        self.jsmem.subarray(addr, addr + len(b)).assign(b)
