+++ spy/llwasm/wasmtime.py
@@ -1,14 +1,5 @@
 """
 A pythonic wrapper around wasmtime.
-
-It is called 'LL' for two reasons:
-
-  - it exposes a low-level view on the code, compared to other wrappers which
-    are more higher level (e.g., the concept of strings doesn't exist, we only
-    have ints, floats and bytes of memory).
-
-  - it's an unused prefix: other prefixes as "Py", "Wasm", "W" etc. would have
-    been very confusing :)
 """
 
 from typing import Any, Optional, Literal
@@ -16,29 +7,17 @@ from typing_extensions import Self
 import py.path
 import wasmtime as wt
 import struct
+from .base import HostModule, LLWasmModuleBase, LLWasmInstanceBase, LLWasmMemoryBase, LLWasmType
 
-LLWasmType = Literal[None, 'void *', 'int32_t', 'int16_t']
 ENGINE = wt.Engine()
 
-class LLWasmModule:
+class LLWasmModule(LLWasmModuleBase):
     f: py.path.local
     mod: wt.Module
 
     def __init__(self, f: py.path.local) -> None:
         self.f = f
         self.mod = wt.Module.from_file(ENGINE, str(f))
-
-    def __repr__(self) -> str:
-        return f'<LLWasmModule {self.f}>'
-
-
-class HostModule:
-    """
-    Base class for host modules.
-
-    Each host module can provide one or more WASM import, used by link().
-    """
-    ll: 'LLWasmInstance' # this attribute is set by LLWasmInstance.__init__
 
 
 def get_linker(
@@ -110,7 +89,7 @@ def get_wasi_config() -> wt.WasiConfig:
     return wasi_config
 
 
-class LLWasmInstance:
+class LLWasmInstance(LLWasmInstanceBase):
     f: py.path.local
     store: wt.Store
     instance: wt.Instance
@@ -160,52 +139,8 @@ class LLWasmInstance:
         assert isinstance(func, wt.Func)
         return func(self.store, *args)
 
-    def read_global(self, name: str, deref: LLWasmType = None) -> Any:
-        r"""
-        Read the given global.
 
-        The semantics is a bit unfortunately: currently, clang always store C
-        globals in linear memory, meaning that the corresponding WASM global
-        contains a *pointer* to the memory.
-
-        clang has plans to support "real" WASM globals: see
-        https://github.com/emscripten-core/emscripten/issues/12793
-
-        If we are reading a scalar, we most likely want to deref the pointer
-        and cast the bytes to the desired type. However, in some cases we are
-        interested in the address, e.g. if the global points to an array or a
-        struct: in that case, we can pass deref=None.
-
-        For example, the following C program:
-            int16_t a = 0xAAAA;
-            int16_t b[] = {0xBBBB, 0xCCCC};
-
-        Produces the following WASM:
-            (global $a i32 (i32.const 1024))
-            (global $b i32 (i32.const 1026))
-            (data $d0 (i32.const 1024) "\aa\aa\bb\bb\cc\cc")
-
-        In this case:
-            read_global('a', deref=None) == 1024
-            read_global('b', deref=None) == 1026
-            read_global('a', deref='i16') == 0xAAAA
-            read_global('b', deref='i16') == 0xBBBB # first item of the array
-        """
-        g = self.get_export(name)
-        assert isinstance(g, wt.Global)
-        addr = g.value(self.store)
-        assert isinstance(addr, int)
-        if deref is None:
-            return addr
-        elif deref == 'int32_t' or deref == 'void *':
-            return self.mem.read_i32(addr)
-        elif deref == 'int16_t':
-            return self.mem.read_i16(addr)
-        else:
-            assert False, f'Unknown type: {deref}'
-
-
-class LLWasmMemory:
+class LLWasmMemory(LLWasmMemoryBase):
     """
     Thin wrapper around wt.Memory
     """
@@ -222,61 +157,5 @@ class LLWasmMemory:
         """
         return self.mem.read(self.store, addr, addr+n)
 
-    def read_i32(self, addr: int) -> int:
-        rawbytes = self.read(addr, 4)
-        return struct.unpack('i', rawbytes)[0]
-
-    def read_i16(self, addr: int) -> int:
-        rawbytes = self.read(addr, 2)
-        return struct.unpack('h', rawbytes)[0]
-
-    def read_i8(self, addr: int) -> int:
-        rawbytes = self.read(addr, 1)
-        return rawbytes[0]
-
-    def read_f64(self, addr: int) -> int:
-        rawbytes = self.read(addr, 8)
-        return struct.unpack('d', rawbytes)[0]
-
-    def read_ptr(self, addr: int) -> tuple[int, int]:
-        """
-        Read a ptr, which we represent as a struct {addr; length }
-        """
-        v_addr = self.read_i32(addr)
-        v_length = self.read_i32(addr+4)
-        return v_addr, v_length
-
-    def read_cstr(self, addr: int) -> bytearray:
-        """
-        Read the NULL-terminated string starting at addr.
-
-        WARNING: this is inefficient because it creates a temporary bytearray
-        for every char To be faster we probably need to bypass wasmtime and
-        access directly the ctypes-based view of the memory.
-        """
-        n = 0
-        while self.read_i8(addr + n) != 0:
-            n += 1
-        return self.read(addr, n)
-
     def write(self, addr: int, b: bytes) -> None:
         self.mem.write(self.store, b, addr)
-
-    def write_i32(self, addr: int, v: int) -> None:
-        self.write(addr, struct.pack('i', v))
-
-    def write_i16(self, addr: int, v: int) -> None:
-        self.write(addr, struct.pack('h', v))
-
-    def write_i8(self, addr: int, v: int) -> None:
-        self.write(addr, struct.pack('b', v))
-
-    def write_f64(self, addr: int, v: float) -> None:
-        self.write(addr, struct.pack('d', v))
-
-    def write_ptr(self, addr: int, v_addr: int, v_length: int) -> None:
-        """
-        Write a ptr { addr; length } to the given addr
-        """
-        self.write_i32(addr, v_addr)
-        self.write_i32(addr+4, v_length)
